+++ src/js/views/controls/topbar/MissionLi.tsx
@@ -4,16 +4,22 @@ import Portal from "../../../core/base/view/Portal";
 import MissionContextMenu, {Exercise} from "./MissionContextMenu";
 import CaskProgress from "./CaskProgress";
 import classNames from "classnames";
+import { findLastIndex, lowerFirst } from "lodash";
 
 require('../../../../css/blocks/menu/mission.less');
 require('../../../../css/blocks/menu/combolist.less');
 
 export interface MissionProgress {
-    exercise_idx: number;
-    exercise_idx_last: number;
-    exercise_idx_passed: number;
-    exercise_idx_passed_max: number;
-    exercise_idx_available: number;
+    id: number;
+    idx_exercise_current: number;
+    idx_exercise_passed: number;
+    is_passed: boolean;
+    exercises: ExerciseProgress[];
+}
+
+interface ExerciseProgress {
+    id: number;
+    is_passed: boolean;
 }
 
 interface MissionLiProps {
@@ -153,31 +159,32 @@ export default class MissionLi extends React.Component<MissionLiProps, MissionLi
     }
 
     render() {
-        const progress = this.props.progress;
-        const exercise_num_total = progress.exercise_idx_last + 1;
-        let exercise_num_current = 0;
-        let detached = true;
-
-        if (progress.exercise_idx_passed_max === progress.exercise_idx_passed) {
-            // Synced indices: user follows mission without switching to previous exercises
-            exercise_num_current = progress.exercise_idx_passed + 1;
-            detached = false;
-        } else if (progress.exercise_idx_available < progress.exercise_idx) {
-            // Admin switching: show progress as if user follows without skipping
-            exercise_num_current = progress.exercise_idx + 1;
-        } else {
-            // User switching: do not show that current exercise was passed
-            exercise_num_current = progress.exercise_idx;
+        if (!this.props.progress) {
+            throw Error("Progress is empty");
         }
 
-        let perc_pass  = 100 * exercise_num_current / exercise_num_total;
-        let perc_avail = 100 * (progress.exercise_idx_passed_max + 1) / exercise_num_total;
+        const progress = this.props.progress;
+        const exercise_num_total = progress.exercises.length;
+
+        // are indices synced or not
+        let detached = true;
+
+        const idx_exercise_passed_max = findLastIndex(progress.exercises, e => e.is_passed),
+              idx_exercise_last = exercise_num_total - 1;
+
+        if (idx_exercise_passed_max === progress.idx_exercise_current - 1) {
+            // Synced indices: user follows mission without switching to previous exercises
+            detached = false;
+        }
+
+        let perc_pass  = 100 * (progress.idx_exercise_passed + 1) / exercise_num_total;
+        let perc_avail = 100 * (idx_exercise_passed_max + 1) / exercise_num_total;
 
         const {is_current} = this.props;
 
         const klasses_pager__item = classNames({
             'pager__item': true,
-            'pager__item_starred': progress.exercise_idx_passed == progress.exercise_idx_last,
+            'pager__item_starred': idx_exercise_passed_max == idx_exercise_last,
             'pager__item_active': true,
             'pager__item_current': this.props.is_current,
         });
@@ -192,29 +199,38 @@ export default class MissionLi extends React.Component<MissionLiProps, MissionLi
             'cask__content_disposable': is_current
         });
 
-        const klasses_cask_content_alt = classNames({
+        const klasses_cask_content_icon = classNames({
             'cask__content': true,
             'cask__content_undisposable': is_current
         });
 
-        const is_simple_force = progress.exercise_idx_last == 0;
+        // do not render animation elements for casks
+        const is_simple_force = exercise_num_total === 1;
 
         return (
             <li className={klasses_pager__item}>
                 <div className={klasses_cask}>
-                    {detached ? <CaskProgress is_simple_force={is_simple_force}
-                                              percent={perc_avail} is_simple={!is_current}
+                    {/* Passed progress indicator */}
+                    {detached ? <CaskProgress no_animation={false && is_simple_force}
+                                              // animation for the current mission is disabled
+                                              is_animated={!is_current}
+                                              percent={perc_avail} 
                                               style_fg={is_current ? 'primary-weak' : 'primary-weak'}
                                               style_bg={is_current ? 'warning-weak' : null}
                                 />
                               : null
                     }
-                    <CaskProgress is_simple_force={is_simple_force}
-                                  percent={perc_pass} is_simple={detached || !is_current}
+
+                    {/* Current progress indicator */}
+                    <CaskProgress no_animation={false && is_simple_force}
+                                  // animation for the current mission is enabled only when detached
+                                  is_animated={detached || !is_current}
+                                  percent={perc_pass} 
                                   style_fg={is_current ? 'success' : 'success-weak'}
                                   style_bg={is_current ? 'warning-weak' : null}
                     />
 
+                    {/* Main content (mission number) */}
                     <div ref={this.ref_root}
                          className={klasses_cask__content_main}
                          onClick={this.handleMissionClick}
@@ -224,8 +240,9 @@ export default class MissionLi extends React.Component<MissionLiProps, MissionLi
                         {this.props.index + 1}
                     </div>
 
+                    {/* Action icon */}
                     {this.props.is_current ?
-                        <div className={klasses_cask_content_alt}
+                        <div className={klasses_cask_content_icon}
                              onClick={this.handleMissionClick}
                              onContextMenu={this.handleContextMenu}
                         >
@@ -245,7 +262,7 @@ export default class MissionLi extends React.Component<MissionLiProps, MissionLi
 
                                             title={this.props.title}
                                             description={this.props.description}
-                                            current_exercise_idx={this.props.progress.exercise_idx}
+                                            current_exercise_idx={this.props.progress.idx_exercise_current}
                                             admin_url_prefix={this.props.admin_url_prefix}
                                             on_exercise_select={this.props.on_exercise_select}
                         />
@@ -257,25 +274,30 @@ export default class MissionLi extends React.Component<MissionLiProps, MissionLi
 
     private getActiveIcon() {
         if (this.isProgressDetached()) {
-            // On detached currently active mission, suggest the user to
+            // If currently active mission is detached, suggest the user to
             // forward to the last exercise available in the mission
             return <i className="fa fa-fast-forward" aria-hidden="true"/>
         } else {
-            // On synced currently active mission, suggest the user to
+            // If currently active mission is synced, suggest the user to
             // restart the mission from the first exercise
             return <i className="fa fa-redo-alt" aria-hidden="true"/>
         }
     }
 
     /**
-     * Determine whether user follows the mission' exercises consistently.
+     * Determines whether user follows the mission' exercises consistently.
      * If user switches manually, this method returns `false`.
      *
      * @private
      */
     private isProgressDetached() {
         const progress = this.props.progress;
-        return progress.exercise_idx_available !== progress.exercise_idx;
+
+        if (progress.exercises.length === 1) return false;
+
+        const idx_exercise_passed_max = findLastIndex(progress.exercises, e => e.is_passed);
+
+        return progress.idx_exercise_current !== idx_exercise_passed_max + 1;
     }
 }